Remove architecture specific implementations of `fma` and instead use `std::fma` (#87)
Bench: 855318

+++ src/eval/value_network.hpp
@@ -14,7 +14,7 @@ constexpr usize L1_SIZE = 3072;
 constexpr usize L2_SIZE = 16;
 constexpr usize L3_SIZE = 128;
 constexpr usize VECTOR_SIZE = util::NATIVE_SIZE<i16>;
-constexpr usize L2_REG_SIZE = std::min(util::NATIVE_SIZE<f32>, L2_SIZE);
+constexpr usize L2_REG_SIZE = std::min(util::NATIVE_SIZE<i16>, L2_SIZE);
 constexpr usize L3_REG_SIZE = std::min(util::NATIVE_SIZE<f32>, L3_SIZE);
 
 constexpr i16 EVAL_SCALE = 400;
